refactor: SearchHomeViewController Section Layout 함수 분리
- SectionLayoutFactory 추가
- createCompositionalLayout 함수 static 하게 변경

ref #5
