+++ iOS/AirbnbApp/AirbnbApp/Source/Present/SearchHomeViewController.swift
@@ -16,7 +16,7 @@ class SearchHomeViewController: UIViewController {
     private let searchBarDelegate = DestinationSearchBarDelegate()
     
     private lazy var destinationCollectionView: UICollectionView = {
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: createCompositionalLayout())
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: SectionLayoutFactory.createCompositionalLayout())
         
         collectionView.register(HeroImageViewCell.self, forCellWithReuseIdentifier: HeroImageViewCell.identifier)
         collectionView.register(NearDestinationViewCell.self, forCellWithReuseIdentifier: NearDestinationViewCell.identifier)
@@ -120,138 +120,4 @@ private extension SearchHomeViewController {
         }
     }
     
-    func createCompositionalLayout() -> UICollectionViewCompositionalLayout {
-        return UICollectionViewCompositionalLayout { (sectionNumber, _) -> NSCollectionLayoutSection? in
-            guard let sectionKind = DestinationCollectionViewSection(rawValue: sectionNumber) else { return nil }
-            let section: NSCollectionLayoutSection
-            switch sectionKind {
-            case .image:
-                section = self.heroImageLayoutSection()
-            case .nearby:
-                section = self.nearDestinationLayoutSection()
-            case .theme:
-                section = self.themeLayoutSection()
-            }
-            return section
-        }
-    }
-    
-    func heroImageLayoutSection() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(1),
-            heightDimension: .fractionalHeight(1)
-        )
-        
-        let item = NSCollectionLayoutItem(
-            layoutSize: itemSize)
-        item.contentInsets.bottom = 15
-        
-        let groupSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(1.0),
-            heightDimension: .fractionalWidth(1.0)
-        )
-        
-        let group = NSCollectionLayoutGroup.vertical(
-            layoutSize: groupSize,
-            subitems: [item])
-        group.contentInsets = .init(
-            top: 0,
-            leading: 0,
-            bottom: 0,
-            trailing: 0
-        )
-        
-        let section = NSCollectionLayoutSection(group: group)
-        
-        section.boundarySupplementaryItems = [
-            NSCollectionLayoutBoundarySupplementaryItem(
-                layoutSize: .init(
-                    widthDimension: .fractionalWidth(1),
-                    heightDimension: .estimated(44)),
-                elementKind: UICollectionView.elementKindSectionHeader,
-                alignment: .topLeading)
-        ]
-        
-        return section
-        
-    }
-    
-    func nearDestinationLayoutSection() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(0.7),
-            heightDimension: .fractionalHeight(0.2)
-        )
-        
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = .init(
-            top: 0,
-            leading: 0,
-            bottom: 15,
-            trailing: 0
-        )
-        
-        let groupSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(0.9),
-            heightDimension: .fractionalHeight(0.25)
-        )
-        
-        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitem: item, count: 2)
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets.leading = 15
-        
-        section.boundarySupplementaryItems = [
-            NSCollectionLayoutBoundarySupplementaryItem(
-                layoutSize: .init(
-                    widthDimension: .fractionalWidth(1),
-                    heightDimension: .absolute(56)),
-                elementKind: UICollectionView.elementKindSectionHeader,
-                alignment: .topLeading)
-        ]
-        section.orthogonalScrollingBehavior = .continuous
-        return section
-    }
-    
-    func themeLayoutSection() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(1),
-            heightDimension: .fractionalHeight(1)
-        )
-        
-        let item = NSCollectionLayoutItem(
-            layoutSize: itemSize)
-        item.contentInsets.bottom = 15
-        
-        let groupSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalWidth(0.8),
-            heightDimension: .fractionalWidth(0.8)
-        )
-        
-        let group = NSCollectionLayoutGroup.horizontal(
-            layoutSize: groupSize,
-            subitems: [item]
-        )
-        group.contentInsets = .init(
-            top: 0,
-            leading: 0,
-            bottom: 15,
-            trailing: 16
-        )
-        
-        let section = NSCollectionLayoutSection(group: group)
-        
-        section.orthogonalScrollingBehavior = .continuous
-        section.contentInsets.leading = 15
-        
-        section.boundarySupplementaryItems = [
-            NSCollectionLayoutBoundarySupplementaryItem(
-                layoutSize: .init(
-                    widthDimension: .fractionalWidth(1),
-                    heightDimension: .absolute(56)),
-                elementKind: UICollectionView.elementKindSectionHeader,
-                alignment: .topLeading)
-        ]
-        
-        return section
-    }
 }
