+++ iOS/AirbnbApp/AirbnbApp/Source/Present/SectionLayoutFactory.swift
@@ -1,0 +1,145 @@
+//
+//  SectionLayoutFactory.swift
+//  AirbnbApp
+//
+//  Created by dale on 2022/05/25.
+//
+
+import UIKit
+
+enum SectionLayoutFactory {
+    static func createCompositionalLayout() -> UICollectionViewCompositionalLayout {
+        return UICollectionViewCompositionalLayout { (sectionNumber, _) -> NSCollectionLayoutSection? in
+            guard let sectionKind = DestinationCollectionViewSection(rawValue: sectionNumber) else { return nil }
+            let section: NSCollectionLayoutSection
+            switch sectionKind {
+            case .image:
+                section = heroImageLayoutSection()
+            case .nearby:
+                section = nearDestinationLayoutSection()
+            case .theme:
+                section = themeLayoutSection()
+            }
+            return section
+        }
+    }
+    
+    static func heroImageLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1),
+            heightDimension: .fractionalHeight(1)
+        )
+        
+        let item = NSCollectionLayoutItem(
+            layoutSize: itemSize)
+        item.contentInsets.bottom = 15
+        
+        let groupSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1.0),
+            heightDimension: .fractionalWidth(1.0)
+        )
+        
+        let group = NSCollectionLayoutGroup.vertical(
+            layoutSize: groupSize,
+            subitems: [item])
+        group.contentInsets = .init(
+            top: 0,
+            leading: 0,
+            bottom: 0,
+            trailing: 0
+        )
+        
+        let section = NSCollectionLayoutSection(group: group)
+        
+        section.boundarySupplementaryItems = [
+            NSCollectionLayoutBoundarySupplementaryItem(
+                layoutSize: .init(
+                    widthDimension: .fractionalWidth(1),
+                    heightDimension: .estimated(44)),
+                elementKind: UICollectionView.elementKindSectionHeader,
+                alignment: .topLeading)
+        ]
+        
+        return section
+        
+    }
+    
+    static func nearDestinationLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(0.7),
+            heightDimension: .fractionalHeight(0.2)
+        )
+        
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = .init(
+            top: 0,
+            leading: 0,
+            bottom: 15,
+            trailing: 0
+        )
+        
+        let groupSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(0.9),
+            heightDimension: .fractionalHeight(0.25)
+        )
+        
+        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitem: item, count: 2)
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.contentInsets.leading = 15
+        
+        section.boundarySupplementaryItems = [
+            NSCollectionLayoutBoundarySupplementaryItem(
+                layoutSize: .init(
+                    widthDimension: .fractionalWidth(1),
+                    heightDimension: .absolute(56)),
+                elementKind: UICollectionView.elementKindSectionHeader,
+                alignment: .topLeading)
+        ]
+        section.orthogonalScrollingBehavior = .continuous
+        return section
+    }
+    
+    static func themeLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1),
+            heightDimension: .fractionalHeight(1)
+        )
+        
+        let item = NSCollectionLayoutItem(
+            layoutSize: itemSize)
+        item.contentInsets.bottom = 15
+        
+        let groupSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(0.8),
+            heightDimension: .fractionalWidth(0.8)
+        )
+        
+        let group = NSCollectionLayoutGroup.horizontal(
+            layoutSize: groupSize,
+            subitems: [item]
+        )
+        group.contentInsets = .init(
+            top: 0,
+            leading: 0,
+            bottom: 15,
+            trailing: 16
+        )
+        
+        let section = NSCollectionLayoutSection(group: group)
+        
+        section.orthogonalScrollingBehavior = .continuous
+        section.contentInsets.leading = 15
+        
+        section.boundarySupplementaryItems = [
+            NSCollectionLayoutBoundarySupplementaryItem(
+                layoutSize: .init(
+                    widthDimension: .fractionalWidth(1),
+                    heightDimension: .absolute(56)),
+                elementKind: UICollectionView.elementKindSectionHeader,
+                alignment: .topLeading)
+        ]
+        
+        return section
+    }
+}
